Modified DetailsVC to show the Item name in the navigation bar.
Deleted ItemsHeader.xib.

Modified ItemsVC to display "Homeowner" in the navigation bar with buttons that subsumed the functionality previously present in the ItemsHeader.  Removed references to the ItemsHeader throughout.  Removed the commented out commitEditingStyle code.  Added functionality to change and redisplay the last cell for either "No Items" or "No More Items", depending on the exist of any items.  Removed the toggleEditingMode action.