Added necessary camera use key to Info.plist to allow use of the camera.
Modified DetailsVC to show a light gray background for the image view to make it obvious where the image would be if there was one.

Modified DetailsVC to present an appropriate image picker when the camera button is tapped in the toolbar, displaying the picked/snapped image in the image view and dismissing the image picker when one is received.