Cleaned up DetailsVC to use the dot notation in place of a message pass notation in one spot.
Added ImageVC for the display of an image as a popover or modal image view.

Modified ItemCell with an actionBlock property that can be set to the action to use in showImage.

Modified ItemsVC to provide an actionBlock to display the image as a popover or modal dialog with the button situated over the cell's image is tapped.