Added a Core Data data model with Item and AssetType entities.  Item's class needs to be refactored to support Core Data functionality while maintaining existing functionality.  AssetType's class can simply be generated.
Refactored Item as a Core Data managed object.

Modified ItemStore to provide allAssetTypes in addition to allItems.  Also refactored it substantially to use the Core Data model for persistence in lieu of keyed archiving.

Modified ItemsVC to use the refactored Core Data extension of Item.

Added ImageTransformer to negotiation between the persisted NSData in the Core Data model and the PNG Image representation of Item's thumbnail.

Added AssetTypeVC to display the available Asset Type, the one selected for the current Item, and to allow for choosing one.

Modified DetailsVC to use the refactored Core Data extension of Item.  Also added a button that both shows the selected Asset Type for the current Item, and if selected, presents the aforementioned AssetTypeVC.